+++ CHANGELOG.md
@@ -27,3 +27,4 @@
 - Fix unit tests
 - Fix bug when trying to google translate an empty string
 - Add support for `!foo` placeholders using google translate
+- Show the diff when a change is detected (allowing someone to manually fix it)